Switch IP query from dig to curl
For some reason the DNS approach wasn't working out on
WireGuard+Mullvad. Using curl/HTTP for this is all-around shittier
and less reliable, but such is life. This also makes the IPv4 and IPv6
config options always return the same thing as eachother, but I'm going
to leave them both in place anyways because this is a personal fork and
I want to keep my config largely compatible with master.